Refactor task screens to use TaskListWidget for better code organization and maintainability
