+++ lib/ui/screens/completed_task_screen.dart
@@ -1,14 +1,14 @@
 import 'package:flutter/material.dart';
-import 'package:task_manager_flutter/ui/screens/task_screen.dart';
+import 'package:task_manager_flutter/ui/widgets/task_list_widget.dart';
 
 class CompleteTaskScreen extends StatelessWidget {
   const CompleteTaskScreen({super.key});
 
   @override
   Widget build(BuildContext context) {
-    return TaskScreen(
-      screenStatus: 'Completed',
-      apiLink: '', // Not needed for Firestore
+    return const TaskListWidget(
+      status: 'Completed',
+      title: 'Completed Tasks',
     );
   }
 }
