+++ lib/ui/screens/progress_task_screen.dart
@@ -1,14 +1,14 @@
 import 'package:flutter/material.dart';
-import 'package:task_manager_flutter/ui/screens/task_screen.dart';
+import 'package:task_manager_flutter/ui/widgets/task_list_widget.dart';
 
 class ProgressTaskScreen extends StatelessWidget {
   const ProgressTaskScreen({super.key});
 
   @override
   Widget build(BuildContext context) {
-    return TaskScreen(
-      screenStatus: 'In Progress',
-      apiLink: '', // Not needed for Firestore
+    return const TaskListWidget(
+      status: 'In Progress',
+      title: 'In Progress Tasks',
     );
   }
 }
