+++ lib/ui/widgets/task_list_widget.dart
@@ -1,0 +1,242 @@
+import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:firebase_auth/firebase_auth.dart';
+import 'package:flutter/material.dart';
+import 'package:task_manager_flutter/ui/screens/add_task_screen.dart';
+
+class TaskListWidget extends StatefulWidget {
+  final String status;
+  final String title;
+
+  const TaskListWidget({
+    Key? key,
+    required this.status,
+    required this.title,
+  }) : super(key: key);
+
+  @override
+  State<TaskListWidget> createState() => _TaskListWidgetState();
+}
+
+class _TaskListWidgetState extends State<TaskListWidget> {
+  List<Map<String, dynamic>> _tasks = [];
+  bool _isLoading = true;
+
+  @override
+  void initState() {
+    super.initState();
+    _fetchTasks();
+  }
+
+  Future<void> _fetchTasks() async {
+    setState(() {
+      _isLoading = true;
+    });
+
+    try {
+      final userId = FirebaseAuth.instance.currentUser?.uid;
+      if (userId == null) {
+        throw Exception("User not logged in");
+      }
+
+      final tasksSnapshot = await FirebaseFirestore.instance
+          .collection('users')
+          .doc(userId)
+          .collection('tasks')
+          .where('status', isEqualTo: widget.status)
+          .get();
+
+      setState(() {
+        _tasks = tasksSnapshot.docs.map((doc) => {'id': doc.id, ...doc.data()}).toList();
+      });
+    } catch (e) {
+      ScaffoldMessenger.of(context).showSnackBar(
+        SnackBar(content: Text("Failed to load tasks: ${e.toString()}")),
+      );
+    } finally {
+      setState(() {
+        _isLoading = false;
+      });
+    }
+  }
+
+  Future<void> _deleteTask(String taskId) async {
+    try {
+      final userId = FirebaseAuth.instance.currentUser?.uid;
+      if (userId == null) {
+        throw Exception("User not logged in");
+      }
+
+      await FirebaseFirestore.instance.collection('users').doc(userId).collection('tasks').doc(taskId).delete();
+
+      setState(() {
+        _tasks.removeWhere((task) => task['id'] == taskId);
+      });
+
+      ScaffoldMessenger.of(context).showSnackBar(
+        const SnackBar(content: Text("Task deleted successfully!")),
+      );
+    } catch (e) {
+      ScaffoldMessenger.of(context).showSnackBar(
+        SnackBar(content: Text("Failed to delete task: ${e.toString()}")),
+      );
+    }
+  }
+
+  void _showEditTaskBottomSheet(Map<String, dynamic> task) {
+    final TextEditingController titleController = TextEditingController(text: task['title']);
+    final TextEditingController descriptionController = TextEditingController(text: task['description']);
+    String selectedStatus = task['status'];
+
+    showModalBottomSheet(
+      context: context,
+      shape: RoundedRectangleBorder(
+        borderRadius: BorderRadius.circular(20),
+      ),
+      builder: (context) {
+        return Padding(
+          padding: const EdgeInsets.all(16.0),
+          child: Column(
+            mainAxisSize: MainAxisSize.min,
+            crossAxisAlignment: CrossAxisAlignment.start,
+            children: [
+              const Text(
+                "Edit Task",
+                style: TextStyle(fontSize: 20, fontWeight: FontWeight.bold),
+              ),
+              const SizedBox(height: 16),
+              TextField(
+                controller: titleController,
+                decoration: InputDecoration(
+                  labelText: "Title",
+                  border: OutlineInputBorder(
+                    borderRadius: BorderRadius.circular(10),
+                  ),
+                ),
+              ),
+              const SizedBox(height: 16),
+              TextField(
+                controller: descriptionController,
+                maxLines: 3,
+                decoration: InputDecoration(
+                  labelText: "Description",
+                  border: OutlineInputBorder(
+                    borderRadius: BorderRadius.circular(10),
+                  ),
+                ),
+              ),
+              const SizedBox(height: 16),
+              DropdownButtonFormField<String>(
+                value: selectedStatus,
+                items: ['New', 'In Progress', 'Completed', 'Cancelled']
+                    .map((status) => DropdownMenuItem(
+                          value: status,
+                          child: Text(status),
+                        ))
+                    .toList(),
+                onChanged: (value) {
+                  if (value != null) {
+                    selectedStatus = value;
+                  }
+                },
+                decoration: InputDecoration(
+                  labelText: "Status",
+                  border: OutlineInputBorder(
+                    borderRadius: BorderRadius.circular(10),
+                  ),
+                ),
+              ),
+              const SizedBox(height: 16),
+              ElevatedButton(
+                onPressed: () async {
+                  try {
+                    final userId = FirebaseAuth.instance.currentUser?.uid;
+                    if (userId == null) {
+                      throw Exception("User not logged in");
+                    }
+
+                    await FirebaseFirestore.instance
+                        .collection('users')
+                        .doc(userId)
+                        .collection('tasks')
+                        .doc(task['id'])
+                        .update({
+                      'title': titleController.text.trim(),
+                      'description': descriptionController.text.trim(),
+                      'status': selectedStatus,
+                    });
+
+                    Navigator.pop(context);
+                    _fetchTasks();
+
+                    ScaffoldMessenger.of(context).showSnackBar(
+                      const SnackBar(content: Text("Task updated successfully!")),
+                    );
+                  } catch (e) {
+                    ScaffoldMessenger.of(context).showSnackBar(
+                      SnackBar(content: Text("Failed to update task: ${e.toString()}")),
+                    );
+                  }
+                },
+                child: const Text("Save Changes"),
+              ),
+            ],
+          ),
+        );
+      },
+    );
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(title: Text(widget.title)),
+      body: RefreshIndicator(
+        onRefresh: () async {
+          await _fetchTasks();
+        },
+        child: _isLoading
+            ? const Center(child: CircularProgressIndicator())
+            : _tasks.isEmpty
+                ? const Center(child: Text("No tasks found"))
+                : ListView.builder(
+                    itemCount: _tasks.length,
+                    itemBuilder: (context, index) {
+                      final task = _tasks[index];
+                      return Card(
+                        margin: const EdgeInsets.symmetric(horizontal: 16, vertical: 8),
+                        shape: RoundedRectangleBorder(
+                          borderRadius: BorderRadius.circular(10),
+                        ),
+                        child: ListTile(
+                          title: Text(task['title'] ?? 'Untitled'),
+                          subtitle: Text(task['description'] ?? 'No description'),
+                          trailing: Row(
+                            mainAxisSize: MainAxisSize.min,
+                            children: [
+                              IconButton(
+                                icon: const Icon(Icons.edit, color: Colors.blue),
+                                onPressed: () => _showEditTaskBottomSheet(task),
+                              ),
+                              IconButton(
+                                icon: const Icon(Icons.delete, color: Colors.red),
+                                onPressed: () => _deleteTask(task['id']),
+                              ),
+                            ],
+                          ),
+                        ),
+                      );
+                    },
+                  ),
+      ),
+      floatingActionButton: FloatingActionButton(
+        onPressed: () {
+          Navigator.push(
+            context,
+            MaterialPageRoute(builder: (context) => const AddTaskScreen()),
+          ).then((_) => _fetchTasks()); // Refresh tasks when returning from add screen
+        },
+        child: const Icon(Icons.add),
+      ),
+    );
+  }
+}
